refactor(frontend): extract common query to hook
